+++ dhall/build.rs
@@ -15,7 +15,6 @@ fn convert_abnf_to_pest() -> std::io::Result<()> {
 
     let mut data = read_to_string(abnf_path)?;
     data.push('\n');
-    let data = data.replace('∀', ""); // TODO: waiting for abnf 0.6.1
 
     let mut rules = abnf_to_pest::parse_abnf(&data)?;
     for line in BufReader::new(File::open(visibility_path)?).lines() {
@@ -95,7 +94,7 @@ fn convert_abnf_to_pest() -> std::io::Result<()> {
             bool_or |
             import_alt
         }}
-        operator_expression = {{ with_expression ~ (whsp ~ operator ~ whsp ~ with_expression)* }}
+        operator_expression = {{ application_expression ~ (whsp ~ operator ~ whsp ~ application_expression)* }}
     "##
     )?;